comdlg32/tests: Fix itemdlg tests on Windows.
The wine testbot runs tests with C:\Users\Public\Documents as the
working directory. Itemdlg tests create test files in the working
directory and then try to select them for opening in an IFileOpenDialog.
The dialog has its SetFolder method called to point it to the current
working directory before showing the dialog, in order for the test code
to find the test files it had created there just before opening the
dialog.

On Windows, this failed because calling the dialog's SetFolder method
with C:\Users\Public\Documents doesn't work as expected. The dialog
points to the C:\Users\[CURRENT_USER]\Documents directory instead.
Subsequently, the test would be unable to select the test file because
it doesn't exist in that directory. So the dialog wouldn't accept the
file and close itself to finish the test, but show an error message
instead, resulting in a test timeout.

To fix this, set the current working directory to the temporary
directory before calling individual itemdlg test_* functions.