advapi32/tests: Enumerate the services using the Unicode API.
The ANSI one is broken in some UTF-8 locales which prevents most of the
tests from running. So use the Unicode API for most tests and just
confirm that it is consistent with the ANSI one instead of doing the
reverse.

Wine-Bug: https://bugs.winehq.org//show_bug.cgi?id=52895