taskschd/tests: Deleting tasks & folders requires elevated privileges on Windows 7.
This means skipping any test that requires creating a folder as the
test would be unable to clean up after itself.
Add a test to show that there is the same issue with tasks by creating
one in the root folder.

Wine-Bug: https://bugs.winehq.org//show_bug.cgi?id=54106