wined3d: Sync depth stencils to both early and late depth test.
We don't know which test the next draw will use when we place the
barrier. We may be able to figure this out, but it is not trivial.

This avoids a read-after-write hazard after the next patch. The layout
transition from shader-read-only to depth stencil needs to be finished
before the early depth test in a follow-up draw if this draw uses early
depth.