wldap32: Implement setting LDAP_OPT_SSL and add tests.
Windows has the LDAP_OPT_SSL option to turn SSL on or off after calling
ldap_init but before connecting to the server. OpenLDAP doesn't have
that option and instead expects a fully SSL connection to be requested
by passing one or more "ldaps://" URIs to ldap_init. However, OpenLDAP
also has an LDAP_OPT_URI option, which Windows does not have, that can
be used to change the URIs before connecting. And OpenLDAP already takes
care of converting "ldap://" or "ldaps://" to lowercase, so all we have
to do is find and replace that exact string in each URI in the list.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=54727