wininet/tests: Dump the state string if it is not as expected.
Also set up the buffer so dumping it is safe enough if the string is
not null-terminated.
And fix the failure messages since the tested value is the string length
in characters, not the buffer size in bytes.

(cherry picked from commit f0b9f257c486f4ba1d6f7a1f7469bc3ed0cbfd38)