winewayland.drv: Basic support for Windows monitor positioning.
Use the xdg-output-unstable-v1 protocol to get the positions of the
Wayland outputs in the compositor logical space, and use this
information to arrange the monitors in the Windows virtual screen space.
For now we assume that the logical positions match the physical pixel
positions (i.e., the outputs are not scaled), a deficiency which will be
addressed in an upcoming commit.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>