mfplat/tests: Independently test that locking flags are ignored for D3D9 buffers.
In the previous test the same address was always returned for data,
hinting that the tests might have been succeding just because of some
aliasing reason. The new tests show that this is not the case:
even when transferring data through another throw-away surface,
changes are seen and recorded independently of the locking flags.