oledb32: Put provider initialization properties in a single DBPROP_INIT_PROVIDERSTRING.
Put provider initialization properties in a single DBPROP_INIT_PROVIDERSTRING for IDBProperties::SetProperties().
Otherwise, according to MSDN for IDBProperties::SetProperties(), if the same property is specified
more than once, the value used is provider-specific. So it won't be consider as an initialization property.

Fix PowerPivot for Excel not connecting to embedded data sources.