comdlg32: Set a size for toolbar buttons of the file access dialog.
The default size is too small for HiDPI, so set the same height as
the look-in combo box to it. And since in owner draw mode, the height
of combo box is calculated and updated when drawing items, this patch
set button height for toolbar from FILEDLG95_LOOKIN_DrawItem() too.