qcap/tests: Remove duplicated tests for allocator identity.
We test once above that the source does not use the existing allocator, but then
implicitly release the allocator we are testing against (when the source assigns
a new allocator to the sink). As a result, these ok() statements were testing
that newly allocated allocators had a different address than the original
allocator, which of course sometimes fails.

Since we don't need to perform this test more than once anyway, just get rid of
the offending ok() statements.