winegstreamer: Only warn on wg_transform input buffer push errors.
Some GStreamer plugins such as openh264 return spurious errors when
running the tests. They pass the tests successfull if we simply ignore
them.

It does not make much difference with returning the error, as they are
not supposed to happen anyway, and most of the time the MFT clients
don't expect or handle errors.