ntdll: Do not throw deadlock exceptions for Wine critical sections.
This is actually a public debugging facility, but currently applies exclusively
to Wine internal locks. While we could expand this feature to all critical
sections, no application is known to depend on it, and the fact that it is a
debugging feature means it is not very likely we will ever find one.

As far as Wine is concerned, I believe this feature is unused. I have not heard
any mention of it in at least five years.

Moreover, we have a different tool for debugging critical section timeouts,
namely, the ERR messages in the same function. These messages provide just as
much information while being less destructive to the process state.