winevulkan: Include function name in asserts.
It is really useful to see at a glance what function is
segfaulting/crashing here when debugging games/apps.

This code is generated, and when generating WineVulkan against different
Vulkan spec versions, etc it may be misleading that the Line: referred
to by the assert dialog does not match what is generated locally on
a different machine for the same Wine version/commit.

Signed-off-by: Joshua Ashton <joshua@froggi.es>