dxgi: Consider vk_format a frontend field.
This is an oversight in eec9c3a2f63555604dd6e357bb3280e82e14ff5b:
vk_format specifies the format used for the frontend resources. The backend
images could use a different format, depending on what the swapchain
supports.

Also, introduce the terms "frontend" and "backend" for the two sides
of the swapchain, which are probably easier to understand.