opengl32: Copy to the wow64 buffer even for write maps.
GL_MAP_WRITE_BIT does not mean that the buffer will be filled, unless an
INVALIDATE bit is explicitly set. The application is free to partially update
the buffer, even if it does not read from it.

Thanks to Aida Jonikienė for doing most of the debugging here.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=55045