wined3d: Trigger constant update when an SRV requiring NPOT fixup is bound.
Ported from state_sampler().

This is currently relevant to the case where a d3d11 device is created with a
9.x feature level—this will have resources bound via SRVs [and will therefore
go through wined3d_context_gl_bind_shader_resources() rather than
state_sampler()] but need not support ARB_texture_non_power_of_two.

The NPOT non-requirement is explicitly called out in the d3d11 documentation,
and because we need some degree of emulation for conditional NPOT textures, we
need to hook that up for feature level 9.x d3d11 devices as well.

Looking forward, d3d 1-9 will be normalized to creating internal SRVs and
binding those through the d3d10+ path, so this will also be necessary for that
reason.