wined3d: Remove BO users from the list when invalidating them.
This makes it easier to invalidate in the case where multiple resources share
the same BO, which in turn is necessary to implement copy-on-write semantics for
BOs.

This is also necessary if we ever want to evict resources which have views,
although it's not clear if this will ever be necessary. If nothing else, though,
it removes that implicit dependency.