winegstreamer: Leave media source critical section before unlocking workqueue.
Today, if media_source_Shutdown is called around the same time
as a work item is put on the async_commands_queue, we end up in
a deadlock if Shutdown enters media source's cs first, as it
waits for the queue's callback to finish, which, in turn,
waits for the object's cs to be released.

To avoid this leave the cs, before unlocking the workqueue,
to let any callback on the queue finish running.

Signed-off-by: Bernhard Kölbl <bkoelbl@codeweavers.com>