d3d9/tests: Request a 3 mip level cube texture in test_getdc().
If we let d3d9 decide we get a single-level texture on HW that does not
support mipmapped cubes, e.g. r200 GPUs. By explicitly requesting the 3
levels we need we'll skip the test.

The test demonstrates that only one subresource can be locked / have a
DC, so it makes sense to show that this applies to the whole texture and
not just one level.