winewayland.drv: Remove wayland_surface destruction double locking.
Ignoring the possibility of HWND recycling allows us to use a simpler
scheme to ensure valid access to the wayland_surface associated with an
xdg_surface during event handling. The scheme involves setting HWND as
the xdg_surface user data and using that to get to the wayland_surface.

The prerequisite for this code to be correct is that the wayland surface
destruction for a HWND must be performed under the wayland_win_data
mutex.