winewayland.drv: Handle client-initiated user-driven window move.
When we are notified by Wine core that the user wants to interactively
move the window, forward the request to compositor, so it can take over
and move the window according to user input.

Note that the move does *not* affect the position of the window in the
Wine virtual screen space.