winewayland.drv: Handle client-initiated user-driven window resize.
When we are notified by Wine core that the user wants to interactively
resize the window, forward the request to compositor, so it can take
over and resize the window according to user input.

In the process of the interactive resize we receive a stream of
xdg_toplevel configure events informing us about the updated window
sizes.