winewayland.drv: Handle application-initiated fullscreen state.
Detect when an window wants to become fullscreen and request the Wayland
fullscreen state from the compositor. The fullscreen state for windows
is inferred by their size and position relative to the monitor that
contains their largest region.

Since there is no explicit fullscreen state for windows, compositor
initiated fullscreen requests may lead to a size change, but will not
directly affect other aspects of the window state.