winewayland.drv: Use surface geometry to satisfy state size constraints.
The maximized and fullscreen states disallow surface contents that are
larger than the configured size. However, if we have a larger surface,
and only part of the surface contains useful client data, we can inform
the compositor using xdg_surface_set_geometry, so that only the
specified region is considered for compositor-side positioning and
state constraints.

We use the geometry to support the presentation of windows that insist
on particular positions and sizes which may not match the compositor's
requested size (e.g., fullscreen but may overshoot a few pixels outside
the monitor boundary).