winmm: Reimplement joystick axis mapping.
This should more closely match the algorithm used by native, discovered through
extensive manual testing.

This does not include any automated tests, both because I could not get winmm to
recognize a second plugged joystick, and (as I eventually discovered while
testing) the number of tests that would be necessary to conclusively prove this
algorithm turned out to be impractical.