advapi32: Fix rewinding the last path segment in SetSecurityInfo().
In particular, handle the case where an object has no name.

In theory, this should not happen for regular files, but SetSecurityInfo() is
almost certainly not supposed to care about that [i.e. this code probably
belongs in the server, at the very least]. However, fixing that will require
much more work.

While we're at it, rewrite the code to be a little more idiomatic about its
intent.