wined3d: Rename WINED3DUSAGE_PRIVATE to WINED3DUSAGE_CS.
Which is really what it means nowadays: resources created by the
CS. It has mostly implications WRT thread safety.

In particular, the flag doesn't mean that the resource doesn't
participate in memory accounting (that was split into
WINED3DUSAGE_VIDMEM_ACCOUNTING by
723cd0a4ae76ec45320290893c7af7f90b668556), nor that it isn't checked
for Reset purposes (we use parent == NULL for that).