wbemprox: Use correct string id.
Don't use hard coded string indexes as one of the expected string
can be absent, hence decreasing its index; nothing ensures that the
strings are placed in the order of the fields in the smbios structure.
So use, smbios structures' indexes instead.

Wbemprox was also expecting one string too much.

Signed-off-by: Eric Pouech <epouech@codeweavers.com>