winewayland.drv: Ignore spurious size hints.
Due to the asynchronous nature of Wayland events, and the design of the
xdg_toplevel protocol, an xdg configure event may be a reaction to a
request in the application's configuration past, and the size hint may
be out of date. For example:

1. The client commits a 100x100 buffer to a surface.
2.1 The compositor sends xdg configure(100x100, state=activated).
2.2 In the meantime, the client resizes and commits a 50x50 buffer.
3. The client receives the event from (2.1). If we respect the
   size hint, we will resize back to 100x100, although this was
   neither the client's nor the compositor's intention.

To mitigate this we ignore size hints for states that do not
require strict size adherence.