wined3d: Move state objects from state.c to device.c.
The entire rest of state.c is concerned with the GL fixed-function pipeline.

device.c seems the most fitting place for these, since they're associated with
the device and stored in the device. device.c is also one of the few files that
contains client-side code, which these functions are.