wined3d/glsl: Always set WINED3D_SHADER_CONST_FFP_LIGHTS in FFP constant update masks.
It is possible (if somewhat artificial) to use lighting without using ambient
lighting. In this case ffp_light_ambient may not be an active uniform.

We could be more sophisticated here, and add lighting to the mask if we use any
lighting uniform, but this is simpler and probably good enough.